Made some icons.
One is a M1 iMac, another could either be a Mac Studio or a Mac Mini( it's really the users choice), and the last is a Mac Mini, would have made a MacBook, but didn't seem like it would look good the way I had planned on making it since the lid would technically be sticking out of the icons sides if I made it realistic, and making it just shorter than the bottom would kinda just look weird. I also don't want to show any background, so I can't just like put it on a table.